pyproject.toml
[tool.poetry]
name = "pheonix-bot"
version = "0.1.0"
description = "A discord bot using discord.py"
authors = ["MannuVilasara <mannuvilasara@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.3.2"
asyncio = "^3.4.3"
psutil = "^5.9.8"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

